separate adapter
