+++ stwo_cairo_prover/Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "crates/adapted_prover",
+    "crates/adapter",
     "crates/prover",
     "crates/cairo-serialize",
     "crates/cairo-serialize-derive",
@@ -34,6 +35,8 @@ serde = "1.0.207"
 serde_json = "1.0.1"
 stwo_cairo_prover = { path = "crates/prover", version = "~0.1.0" }
 stwo_cairo_utils = { path = "crates/utils", version = "~0.1.0" }
+stwo-cairo-adapter = { path = "crates/adapter" }
+stwo-cairo-common = { path = "crates/common" }
 # TODO(ShaharS): take stwo version from the source repository.
 stwo-prover = { git = "https://github.com/starkware-libs/stwo", rev = "a194fad", features = [
     "parallel",
