+++ stwo_cairo_prover/crates/adapted_prover/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [features]
-std = ["stwo_cairo_prover/std"]
+std = ["stwo-cairo-adapter/std"]
 
 [dependencies]
 clap.workspace = true
@@ -12,6 +12,7 @@ log.workspace = true
 serde_json.workspace = true
 stwo_cairo_prover.workspace = true
 stwo_cairo_utils.workspace = true
+stwo-cairo-adapter.workspace = true
 stwo-prover.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
