+++ stwo_cairo_prover/crates/adapter/src/builtins.rs
@@ -237,8 +237,8 @@ mod builtin_padding {
     use cairo_vm::air_public_input::MemorySegmentAddresses;
     use itertools::Itertools;
 
-    use crate::adapter::builtins::BITWISE_MEMORY_CELLS;
-    use crate::adapter::memory::{value_from_felt252, MemoryBuilder, MemoryValueId};
+    use crate::builtins::BITWISE_MEMORY_CELLS;
+    use crate::memory::{value_from_felt252, MemoryBuilder, MemoryValueId};
 
     pub fn bitwise(segment: &MemorySegmentAddresses, memory: &mut MemoryBuilder) {
         let range = segment.begin_addr..segment.stop_ptr;
@@ -274,12 +274,10 @@ mod test_builtin_segments {
     use rand::rngs::SmallRng;
     use rand::{Rng, SeedableRng};
 
-    use crate::adapter::builtins::BITWISE_MEMORY_CELLS;
-    use crate::adapter::memory::{
-        u128_to_4_limbs, Memory, MemoryBuilder, MemoryConfig, MemoryValue,
-    };
-    use crate::adapter::vm_import::MemoryEntry;
-    use crate::adapter::BuiltinSegments;
+    use crate::builtins::BITWISE_MEMORY_CELLS;
+    use crate::memory::{u128_to_4_limbs, Memory, MemoryBuilder, MemoryConfig, MemoryValue};
+    use crate::vm_import::MemoryEntry;
+    use crate::BuiltinSegments;
 
     /// Asserts that the values at addresses start_addr1 to start_addr1 + segment_length - 1
     /// are equal to values at the addresses start_addr2 to start_addr2 + segment_length - 1.
