+++ stwo_cairo_prover/crates/adapter/src/lib.rs
