+++ stwo_cairo_prover/crates/adapter/src/memory.rs
