+++ stwo_cairo_prover/crates/adapter/src/plain.rs
