+++ stwo_cairo_prover/crates/adapter/src/vm_import/json.rs
