+++ stwo_cairo_prover/crates/prover/src/cairo_air/air.rs
@@ -1,6 +1,7 @@
 use itertools::{chain, Itertools};
 use num_traits::Zero;
 use serde::{Deserialize, Serialize};
+use stwo_cairo_adapter::ProverInput;
 use stwo_cairo_common::prover_types::cpu::CasmState;
 use stwo_cairo_serialize::CairoSerialize;
 use stwo_prover::constraint_framework::{Relation, TraceLocationAllocator, PREPROCESSED_TRACE_IDX};
@@ -31,7 +32,6 @@ use super::range_checks_air::{
     RangeChecksInteractionClaim, RangeChecksInteractionClaimGenerator,
     RangeChecksInteractionElements,
 };
-use crate::adapter::ProverInput;
 use crate::cairo_air::relations;
 use crate::components::memory::{memory_address_to_id, memory_id_to_big};
 use crate::components::{verify_bitwise_xor_9, verify_instruction};
