+++ stwo_cairo_prover/crates/recursion_bridge/Cargo.toml
@@ -8,3 +8,4 @@ stwo_cairo_prover = { path = "../prover" }
 cairo-lang-casm.workspace = true
 stwo-prover.workspace = true
 serde_json.workspace = true
+stwo-cairo-adapter.workspace = true
