+++ stwo_cairo_prover/crates/run_and_prove/Cargo.toml
@@ -9,6 +9,7 @@ log.workspace = true
 serde_json.workspace = true
 stwo_cairo_prover.workspace = true
 stwo_cairo_utils.workspace = true
+stwo-cairo-adapter.workspace = true
 stwo-prover.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
