+++ stwo_cairo_prover/crates/vm_runner/Cargo.toml
@@ -10,5 +10,6 @@ num-traits.workspace = true
 serde_json.workspace = true
 stwo_cairo_prover.workspace = true
 stwo_cairo_utils.workspace = true
+stwo-cairo-adapter.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
