+++ stwo_cairo_prover/crates/vm_runner/src/main.rs
@@ -2,9 +2,9 @@ use std::path::PathBuf;
 use std::process::ExitCode;
 
 use clap::Parser;
-use stwo_cairo_prover::adapter::plain::adapt_finished_runner;
-use stwo_cairo_prover::adapter::vm_import::VmImportError;
-use stwo_cairo_prover::adapter::{ExecutionResources, ProverInput};
+use stwo_cairo_adapter::plain::adapt_finished_runner;
+use stwo_cairo_adapter::vm_import::VmImportError;
+use stwo_cairo_adapter::{ExecutionResources, ProverInput};
 use stwo_cairo_utils::binary_utils::run_binary;
 use stwo_cairo_utils::vm_utils::{run_vm, VmArgs, VmError};
 use thiserror::Error;
